docs: add expire_cache, object_pool, thread_pool design docs

+++ base/concurrent/blocking_queue_test.cc
@@ -43,7 +43,7 @@ TEST(BlockingQueue, single) {
 TEST(BlockingQueue, overwrite) {
   auto q = new BlockingQueue<int>(5);
   for(int i = 0; i < 6; i++) {
-    q->Push((int)i);
+    q->Push(i);
   }
 
   EXPECT_EQ(q->Pop(), 1);
